refactor: change js to module, move events to separated file, create js dir
